+++ a_test.go
@@ -1,10 +1,73 @@
 package immortal
 
 import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"os"
+	"os/signal"
 	"reflect"
 	"runtime"
+	"sync"
 	"testing"
+	"time"
 )
+
+func TestMain(m *testing.M) {
+	switch os.Getenv("GO_WANT_HELPER_PROCESS") {
+	case "sleep10":
+		GWHPsleep10()
+	case "nosleep":
+		GWHPnosleep()
+	case "signalsUDOT":
+		GWHPsignalsUDOT()
+	case "logStdoutStderr":
+		GWHPlogstdoutstderr()
+	default:
+		os.Exit(m.Run())
+	}
+}
+
+// GWHPsleep1 - test function, exit 1 after sleeping 10 seconds
+func GWHPsleep10() {
+	c := make(chan os.Signal)
+	signal.Notify(c, os.Interrupt, os.Kill)
+	select {
+	case <-c:
+		os.Exit(0)
+	case <-time.After(10 * time.Second):
+		os.Exit(1)
+	}
+}
+
+// GWHPnosleep - test function, exit immediately
+func GWHPnosleep() {
+	os.Exit(0)
+}
+
+// GWHPsignalsUDOT - test function for signals up, down, once, terminate
+func GWHPsignalsUDOT() {
+	fmt.Println("5D675098-45D7-4089-A72C-3628713EA5BA")
+	c := make(chan os.Signal)
+	signal.Notify(c, os.Interrupt, os.Kill)
+	select {
+	case <-c:
+		os.Exit(0)
+	case <-time.After(10 * time.Second):
+		os.Exit(1)
+	}
+}
+
+// GWHPlogstdoutstderr - test function for login stdout & stderr
+func GWHPlogstdoutstderr() {
+	for i := 1; i < 5; i++ {
+		if i%3 == 0 {
+			fmt.Fprintf(os.Stderr, "STDERR i: %d\n", i)
+		} else {
+			fmt.Printf("STDOUT i: %d\n", i)
+		}
+	}
+}
 
 /* Test Helpers */
 func expect(t *testing.T, a interface{}, b interface{}) {
@@ -12,4 +75,32 @@ func expect(t *testing.T, a interface{}, b interface{}) {
 	if a != b {
 		t.Fatalf("Expected: %v (type %v)  Got: %v (type %v)  in %s:%d", a, reflect.TypeOf(a), b, reflect.TypeOf(b), fn, line)
 	}
+}
+
+/* prettyPrint */
+func prettyPrint(i interface{}) string {
+	s, _ := json.MarshalIndent(i, "", "\t")
+	return string(s)
+}
+
+// myBuffer - bytes.Buffer thread-safe
+type myBuffer struct {
+	b bytes.Buffer
+	sync.RWMutex
+}
+
+func (b *myBuffer) Read(p []byte) (n int, err error) {
+	b.RLock()
+	defer b.RUnlock()
+	return b.b.Read(p)
+}
+func (b *myBuffer) Write(p []byte) (n int, err error) {
+	b.Lock()
+	defer b.Unlock()
+	return b.b.Write(p)
+}
+func (b *myBuffer) String() string {
+	b.RLock()
+	defer b.RUnlock()
+	return b.b.String()
 }