+++ logger_test.go
@@ -4,7 +4,10 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"strings"
+	"sync"
 	"testing"
+	"time"
 )
 
 func TestNewStderrLogger(t *testing.T) {
@@ -64,4 +67,44 @@ func TestNewLoggerLogger(t *testing.T) {
 	l := NewLogger(cfg, quit)
 	expect(t, true, l != nil)
 	close(quit)
+}
+
+func TestNewLoggerKilled(t *testing.T) {
+	cfg := &Config{
+		Logger: "sleep 300",
+	}
+	quit := make(chan struct{})
+	l := NewLogger(cfg, quit)
+	expect(t, true, l != nil)
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		time.Sleep(time.Second)
+		close(quit)
+		wg.Done()
+	}()
+	wg.Wait()
+	// TODO verify sleep process is killed
+}
+
+func TestNewLoggerRetry(t *testing.T) {
+	var mylog myBuffer
+	log.SetOutput(&mylog)
+	log.SetFlags(0)
+	cfg := &Config{
+		Logger: "false",
+	}
+	quit := make(chan struct{})
+	l := NewLogger(cfg, quit)
+	expect(t, true, l != nil)
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		time.Sleep(2 * time.Second)
+		close(quit)
+		wg.Done()
+	}()
+	wg.Wait()
+	m := strings.Split(mylog.String(), "\n")
+	expect(t, true, len(m) > 1)
 }